requirements.txt
loguru==0.7.3
numpy==2.3.3
Pillow==11.3.0
scikit_learn==1.7.2
torch==2.8.0
torchvision==0.23.0
transformers==4.57.0
